pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dreambench_plus"
version = "0.0.1"
requires-python = ">=3.10"
description = "DreamBench++: A Human-Aligned Benchmark for Personalized Image Generation"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "black",
    "isort",
    "einops",
    "numpy",
    "openai",
    "requests",
    "omegaconf",
    "timm",
    "megfile",
    "wandb",
    "pillow",
    "torch",
    "torchvision",
    "fire",
    "loguru",
    "pendulum",
    "ipdb",
    "krippendorff",
    "streamlit",
    "accelerate",
    "opencv-python",
    "diffusers==0.25.1",
    "transformers==4.35.2",
]

[project.urls]
Homepage = "https://dreambenchplus.github.io"
Repository = "https://github.com/yuangpeng/dreambench_plus"
"Bug Tracker" = "https://github.com/yuangpeng/dreambench_plus/issues"

[tool.setuptools]
packages = ["dreambench_plus"]

[tool.black]
line-length = 128
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 128
